fix: custom components moved into custom folder inside the components folder
