+++ components/custom/index.ts
@@ -1,0 +1,2 @@
+export * from "./dashboard";
+export * from "./profile";
