+++ components/custom/profile/index.ts
