+++ components/shared/index.ts
@@ -6,6 +6,4 @@ export * from "./pagination";
 export * from "./searchbar";
 export * from "./table";
 export * from "./select";
-export * from "./dashboard/index";
-export * from "./profile/index";
 export * from "./floating-input";
